update scripts and gamescene

+++ Assets/Scripts/Health/Health.cs
@@ -5,13 +5,13 @@ public class Health : MonoBehaviour
 {
     [field: SerializeField] public int CurrentValue;
 
-    public event Action<int> ValueChanged;
+    public event Action<float> ValueChanged;
 
     public event Action ValueZeroReached;
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
-        CurrentValue = Mathf.Clamp(CurrentValue - damage, 0, CurrentValue);
+        CurrentValue = Mathf.Clamp(CurrentValue - (int)damage, 0, CurrentValue);
 
         if (CurrentValue == 0)
             ValueZeroReached?.Invoke();
